Removed deep shallow library
Also switched all instances of Int64 to int.

+++ rush/intervalStack.h
@@ -8,8 +8,8 @@
 
 typedef struct stack{
   Interval **intervals;
-  Int64 curSize;
-  Int64 maxSize;
+  int curSize;
+  int maxSize;
   Interval *top; 
 }Stack;
