+++ rush/rush.h
@@ -6,11 +6,8 @@
 #ifndef RUSH
 #define RUSH
 
-/* #include "interface.h" */
-/* #include "minimize.h" */
-
 void initializeProb(int max, int sbjctLen, double gc);
-int isHomologous(Int64 *sl, Int64 n, int sbjctLen);
+int isHomologous(int *sl, int n, int sbjctLen);
 double expVarVarSl(double l, double p);
 double significanceVar(double meanSl, double varSl, int winLen);
 double eVar(double meanSl, double winLen);
